+++ history.md
@@ -4,6 +4,7 @@
 * Modified exports which impacts how the library is instantiated in code
 * Added code for #30 - refresh data per request
 * Replaced internal request abstraction with [`reqlib`](https://github.com/brozeph/reqlib)
+* Added fix for #28 (min_price and max_price params)
 
 # v1.2.2 / 2017.10.26